Simplify JSON format by removing redundant wrapper objects
