+++ src/main/java/org/dita/dost/project/XmlReader.java
@@ -107,7 +107,7 @@ public class XmlReader {
                             .filter(Optional::isPresent)
                             .map(Optional::get)
                             .map(this::toURI)
-                            .map(include -> new ProjectBuilder.ProjectRef(include))
+//                            .map(include -> new ProjectBuilder.ProjectRef(include))
                             .collect(Collectors.toList()),
                     getChildElements(project, ELEM_PUBLICATION).stream()
                             .map(this::readPublication)
@@ -187,23 +187,26 @@ public class XmlReader {
                 getValue(context, ATTR_IDREF),
                 getChildElement(context, ELEM_INPUTS)
                         .map(inputs -> getChildElements(inputs, ELEM_INPUT).stream()
-                                .map(input -> getAttribute(input, ATTR_HREF).map(this::toURI))
+                                .map(input -> getAttribute(input, ATTR_HREF)
+//                                        .map(this::toURI)
+                                )
                                 .filter(Optional::isPresent)
                                 .map(Optional::get)
-                                .map(input -> new ProjectBuilder.Deliverable.Inputs.Input(input))
+//                                .map(input -> new ProjectBuilder.Deliverable.Inputs.Input(input))
+                                .map(this::toURI)
                                 .collect(Collectors.toList())
                         )
-                        .map(inputs -> new ProjectBuilder.Deliverable.Inputs(inputs))
+//                        .map(ProjectBuilder.Deliverable.Inputs::new)
                         .orElse(null),
                 getChildElement(context, ELEM_PROFILE)
                         .map(inputs -> getChildElements(inputs, ELEM_DITAVAL).stream()
                                 .map(input -> getAttribute(input, ATTR_HREF).map(this::toURI))
                                 .filter(Optional::isPresent)
                                 .map(Optional::get)
-                                .map(input -> new ProjectBuilder.Deliverable.Profile.DitaVal(input))
+//                                .map(input -> new ProjectBuilder.Deliverable.Profile.DitaVal(input))
                                 .collect(Collectors.toList())
                         )
-                        .map(inputs -> new ProjectBuilder.Deliverable.Profile(inputs))
+                        .map(ProjectBuilder.Deliverable.Profile::new)
                         .orElse(null)
         );
     }
