+++ src/test/resources/org/dita/dost/project/common.json
@@ -22,18 +22,12 @@
     {
       "name": "Site",
       "id": "site",
-      "inputs": {
-        "inputs": [
-          {
-            "href": "site.ditamap"
-          }
-        ]
-      },
+      "inputs": [
+        "site.ditamap"
+      ],
       "profiles": {
         "ditavals": [
-          {
-            "href": "site.ditaval"
-          }
+          "site.ditaval"
         ]
       }
     }
