+++ src/test/resources/org/dita/dost/project/product.json
@@ -1,8 +1,6 @@
 {
   "includes": [
-    {
-      "href": "common.json"
-    }
+    "common.json"
   ],
   "deliverables": [
     {
