+++ src/test/resources/org/dita/dost/project/recursive.json
@@ -1,7 +1,5 @@
 {
   "includes": [
-    {
-      "href": "recursive.json"
-    }
+    "recursive.json"
   ]
 }
